+++ example/scripts/main.js
@@ -16,23 +16,57 @@ fileInput.onchange = function(event) {
 function loadImage(src) {
   var image = new Image();
   image.src = src;
-  image.onload = drawImage;
+
+  if (typeof window.Worker === 'function') {
+    image.onload = drawImageUsingWorker;
+  } else {
+    image.onload = drawImage;
+  }
 }
 
 function drawImage(event) {
   var image = event.target;
-  var w = image.width;
-  var h = image.height;
+  var width = image.width;
+  var height = image.height;
 
-  canvas.width = canvasSobel.width = w;
-  canvas.height = canvasSobel.height = h;
+  canvas.width = canvasSobel.width = width;
+  canvas.height = canvasSobel.height = height;
 
   context.drawImage(image, 0, 0);
-  var imageData = context.getImageData(0, 0, w, h);
+  var imageData = context.getImageData(0, 0, width, height);
 
   var sobelData = Sobel(imageData);
   var sobelImageData = sobelData.toImageData();
   contextSobel.putImageData(sobelImageData, 0, 0);
+}
+
+function drawImageUsingWorker() {
+  var image = event.target;
+  var width = image.width;
+  var height = image.height;
+
+  canvas.width = canvasSobel.width = width;
+  canvas.height = canvasSobel.height = height;
+
+  context.drawImage(image, 0, 0);
+  var imageData = context.getImageData(0, 0, width, height);
+
+  var ww = new Worker('/example/scripts/webWorker.js');
+
+  ww.onmessage = function(event) {
+    console.log(event.data);
+    if (Object.prototype.toString.call(event.data) === '[object Uint8ClampedArray]') {
+      var sobelData = event.data;
+      var sobelImageData = Sobel.toImageData(sobelData, width, height);
+      contextSobel.putImageData(sobelImageData, 0, 0);
+    }
+  };
+
+  ww.onerror = function(event) {
+    console.error(event);
+  };
+
+  ww.postMessage(imageData);
 }
 
 loadImage('images/valve.png');