+++ sobel.js
@@ -1,6 +1,3 @@
-
-
-
 (function(root) {
   'use strict';
 
@@ -9,8 +6,8 @@
       return new Sobel(imageData);
     }
 
-    var w = imageData.width;
-    var h = imageData.height;
+    var width = imageData.width;
+    var height = imageData.height;
 
     var kernelX = [
       [-1,0,1],
@@ -30,7 +27,7 @@
     function bindPixelAt(data) {
       return function(x, y, i) {
         i = i || 0;
-        return data[((w * y) + x) * 4 + i];
+        return data[((width * y) + x) * 4 + i];
       };
     }
 
@@ -38,8 +35,8 @@
     var pixelAt = bindPixelAt(data);
     var x, y;
 
-    for (y = 0; y < h; y++) {
-      for (x = 0; x < w; x++) {
+    for (y = 0; y < height; y++) {
+      for (x = 0; x < width; x++) {
         var r = pixelAt(x, y, 0);
         var g = pixelAt(x, y, 1);
         var b = pixelAt(x, y, 2);
@@ -51,8 +48,8 @@
 
     pixelAt = bindPixelAt(grayscaleData);
 
-    for (y = 0; y < h; y++) {
-      for (x = 0; x < w; x++) {
+    for (y = 0; y < height; y++) {
+      for (x = 0; x < width; x++) {
         var pixelX = (
             (kernelX[0][0] * pixelAt(x - 1, y - 1)) +
             (kernelX[0][1] * pixelAt(x, y - 1)) +
@@ -77,26 +74,52 @@
           (kernelY[2][2] * pixelAt(x + 1, y + 1))
         );
 
-        var magnitude = Math.sqrt((pixelX * pixelX) + (pixelY * pixelY))>>0;
+        var magnitude = Math.sqrt((pixelX * pixelX) + (pixelY * pixelY))>>>0;
 
         sobelData.push(magnitude, magnitude, magnitude, 255);
       }
     }
 
-    var clampedArray = new Uint8ClampedArray(sobelData);
+    var clampedArray = sobelData;
+
+    if (typeof Uint8ClampedArray === 'function') {
+      clampedArray = new Uint8ClampedArray(sobelData);
+    }
+
     clampedArray.toImageData = function() {
-      try {
-        return new ImageData(clampedArray, w, h);
-      } catch(error) {
-        var canvas = document.createElement('canvas');
-        var context =  canvas.getContext('2d');
-        var imageData = context.createImageData(w, h);
-        imageData.data.set(clampedArray);
-        return imageData;
-      }
+      return Sobel.toImageData(clampedArray, width, height);
     };
 
     return clampedArray;
+  }
+
+  Sobel.toImageData = function toImageData(data, width, height) {
+    if (typeof ImageData === 'function' && Object.prototype.toString.call(data) === '[object Uint16Array]') {
+      return new ImageData(data, width, height);
+    } else {
+      if (typeof window === 'object' && typeof window.document === 'object') {
+        var canvas = document.createElement('canvas');
+
+        if (typeof canvas.getContext === 'function') {
+          var context = canvas.getContext('2d');
+          var imageData = context.createImageData(width, height);
+          imageData.data.set(data);
+          return imageData;
+        } else {
+          return new FakeImageData(data, width, height);
+        }
+      } else {
+        return new FakeImageData(data, width, height);
+      }
+    }
+  };
+
+  function FakeImageData(data, width, height) {
+    return {
+      width: width,
+      height: height,
+      data: data
+    };
   }
 
   if (typeof exports !== 'undefined') {